Author versus Admin views on puzzles
Author and admin views are implemented for pages that show lists of puzzles - index, per-team status, and the map. The author view filters these lists, and also performs an access check before accessing the author/admin pages in the Puzzles folder.

Along the way, I standardized the constructor for EventSpecificPageModel - but I didn't add every single access check we'll need in the other folders.

+++ ServerCore/Pages/Puzzles/Delete.cshtml.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using ServerCore.DataModel;
@@ -9,11 +10,8 @@ namespace ServerCore.Pages.Puzzles
 {
     public class DeleteModel : EventSpecificPageModel
     {
-        private readonly PuzzleServerContext _context;
-
-        public DeleteModel(PuzzleServerContext context)
+        public DeleteModel(PuzzleServerContext serverContext, UserManager<IdentityUser> userManager) : base(serverContext, userManager)
         {
-            _context = context;
         }
 
         [BindProperty]
@@ -28,6 +26,11 @@ namespace ServerCore.Pages.Puzzles
                 return NotFound();
             }
 
+            if (!await CanAdminPuzzle(Puzzle))
+            {
+                return NotFound();
+            }
+
             return Page();
         }
 
@@ -37,6 +40,11 @@ namespace ServerCore.Pages.Puzzles
 
             if (Puzzle != null)
             {
+                if (!await CanAdminPuzzle(Puzzle))
+                {
+                    return NotFound();
+                }
+
                 foreach (ContentFile content in Puzzle.Contents)
                 {
                     await FileManager.DeleteBlobAsync(content.Url);
